Message code is now 1 byte

+++ MessageHandler.h
@@ -9,15 +9,15 @@ public:
 	MessageHandler(uint8_t msgBegin);
 
 	bool Update(); // Check for any messages and return TRUE only if message is completely received
-	uint16_t GetMessageDataLength() const;
+	uint8_t GetMessageDataLength() const;
 	void GetMessage(uint8_t* msg, uint8_t* data);
 
-	void SendMessage(uint8_t msg, uint16_t len, uint8_t* data);
+	void SendMessage(uint8_t msg, uint8_t len, uint8_t* data);
 
 private:
 	uint8_t* _data;
 	uint8_t _msg;
-	uint16_t _len;
+	uint8_t _len;
 	bool _handled;
 	uint8_t _msgBegin;
 };
